refactor(core): move GET /sign-in-settings out of sessionRouter
