+++ packages/core/src/routes/init.ts
@@ -11,6 +11,7 @@ import resourceRoutes from '@/routes/resource';
 import sessionRoutes from '@/routes/session';
 import settingRoutes from '@/routes/setting';
 import signInExperiencesRoutes from '@/routes/sign-in-experience';
+import signInSettingsRoutes from '@/routes/sign-in-settings';
 import statusRoutes from '@/routes/status';
 import swaggerRoutes from '@/routes/swagger';
 
@@ -25,6 +26,7 @@ const createRouters = (provider: Provider) => {
   sessionRoutes(sessionRouter, provider);
 
   const anonymousRouter: AnonymousRouter = new Router();
+  signInSettingsRoutes(anonymousRouter);
   statusRoutes(anonymousRouter);
   swaggerRoutes(anonymousRouter);
